First And Last
haven't used OpenGL, as compared to other similar graphics libraries, is fairly simple. We'll start with setting it up on our system, followed by writing a simple example demonstrating the usage of the library. if you are used to OpenGL game is too easy to play. there is no output because that's just a sample code of the carrom board